Generate and Combine Column Pairs with Nested Data

combn_pair generates combinations of specified columns, combines them, and nests the data. It supports both data.table and data.frame inputs with optional grouping functionality.

combn_pair(
  data,
  cols2bind,
  by = NULL,
  pairs_n = 2,
  sep = "-",
  nest_type = "dt"
)

Arguments

data

data.table or data.frame. The input dataset.

cols2bind

character vector. Column names to be combined into pairs.

by

character vector, optional. Column names for grouping. Default is NULL.

pairs_n

integer, optional. Number of columns in each combination. Must be >= 2 and <= length of cols2bind. Default is 2.

sep

character, optional. Separator used when combining column names. Default is "-".

nest_type

character, optional. Type of nested data to return: "dt" (data.table) or "df" (data.frame). Default is "dt".

Value

A nested data.table:

When by is NULL: returns data.table nested by 'pairs'

When by is specified: returns data.table nested by 'pairs' and grouping variables

Details

The function performs the following main steps:

Validates input parameters

Generates all possible combinations of specified columns

Creates subsets for each combination and renames columns

Merges all subsets and performs nesting

Note

Important considerations:

cols2bind must be a character vector of existing column names

pairs_n must be a positive integer >= 2 and <= number of columns in cols2bind

sep must be a single character string

Grouping variables in by must exist in the data

nest_type must be either "dt" or "df"

Examples

col_names <- c("Sepal.Length", "Sepal.Width", "Petal.Length")
combn_pair(iris, cols2bind = col_names, pairs_n = 2, sep = "&")
#>                        pairs                data
#>                       <char>              <list>
#> 1:  Sepal.Length&Sepal.Width <data.table[150x4]>
#> 2: Sepal.Length&Petal.Length <data.table[150x4]>
#> 3:  Sepal.Width&Petal.Length <data.table[150x4]>
combn_pair(iris, cols2bind = col_names, pairs_n = 2, by = "Species")
#>                        pairs    Species               data
#>                       <char>     <fctr>             <list>
#> 1:  Sepal.Length-Sepal.Width     setosa <data.table[50x3]>
#> 2:  Sepal.Length-Sepal.Width versicolor <data.table[50x3]>
#> 3:  Sepal.Length-Sepal.Width  virginica <data.table[50x3]>
#> 4: Sepal.Length-Petal.Length     setosa <data.table[50x3]>
#> 5: Sepal.Length-Petal.Length versicolor <data.table[50x3]>
#> 6: Sepal.Length-Petal.Length  virginica <data.table[50x3]>
#> 7:  Sepal.Width-Petal.Length     setosa <data.table[50x3]>
#> 8:  Sepal.Width-Petal.Length versicolor <data.table[50x3]>
#> 9:  Sepal.Width-Petal.Length  virginica <data.table[50x3]>