Convert Nested Columns Between Data Frame and Data Table

This function converts a data frame or data table, transforming nested columns to either data frame or data table format while preserving the original data structure.

convert_nested(data, to = c("df", "dt"), nested_cols = NULL)

Arguments

data

A data frame or data table containing nested columns

to

A character string specifying the target format. Options are "df" (data frame) or "dt" (data table). Defaults to "df".

nested_cols

A character vector of column names containing nested data. If NULL, the function automatically detects list columns.

Value

A data frame or data table with nested columns converted to the specified format

Details

The function performs the following operations:

Automatically detects nested columns if not explicitly specified

Converts the entire data structure to the target format (data frame or data table)

Converts nested columns within the specified columns to the target format

Creates a copy of the input data to avoid modifying the original

Note

The function preserves the original data structure

Nested columns are converted individually

Automatic detection of nested columns is based on list-type columns

Examples

# Convert a data frame with nested columns to data table
df_nest1 <- iris |> 
  dplyr::group_nest(Species)
df_nest2 <- iris |>
  dplyr::group_nest(Species) |>
  dplyr::mutate(data2 = purrr::map(data, dplyr::mutate, c=2))
# Convert a data table with specific nested columns to data frame
convert_nested(df_nest1, to = "dt", nested_cols = c("data"))
#>       Species               data
#>        <fctr>             <list>
#> 1:     setosa <data.table[50x4]>
#> 2: versicolor <data.table[50x4]>
#> 3:  virginica <data.table[50x4]>
convert_nested(df_nest2, to = "dt", nested_cols = c("data", "data2"))
#>       Species               data              data2
#>        <fctr>             <list>             <list>
#> 1:     setosa <data.table[50x4]> <data.table[50x5]>
#> 2: versicolor <data.table[50x4]> <data.table[50x5]>
#> 3:  virginica <data.table[50x4]> <data.table[50x5]>
# Convert a data table with nested columns to data frame
dt_nest <- mintyr::w2l_nest(data = iris, cols2l = 1:2, by = "Species")
convert_nested(dt_nest, to = "dt", nested_cols = c("data"))
#>            name    Species               data
#>          <fctr>     <fctr>             <list>
#> 1: Sepal.Length     setosa <data.table[50x3]>
#> 2: Sepal.Length versicolor <data.table[50x3]>
#> 3: Sepal.Length  virginica <data.table[50x3]>
#> 4:  Sepal.Width     setosa <data.table[50x3]>
#> 5:  Sepal.Width versicolor <data.table[50x3]>
#> 6:  Sepal.Width  virginica <data.table[50x3]>