Column to Pair Nested Transformation

A sophisticated data transformation tool for generating column pair combinations and creating nested data structures with advanced configuration options.

c2p_nest(data, cols2bind, by = NULL, pairs_n = 2, sep = "-", nest_type = "dt")

Arguments

data

Input 'data frame' or 'data table'

Must contain valid columns for transformation

Supports multiple data types

cols2bind

'character' vector of column names for pair generation

Specifies target columns for combination

Must be existing columns in the dataset

by

Optional 'character' vector for grouping

Enables hierarchical nested transformations

Supports multi-level aggregation

Default is NULL

pairs_n

'numeric' indicating combination size

Minimum value: 2

Maximum value: Length of cols2bind

Controls column pair complexity

Default is 2

sep

'character' separator for pair naming

Used in generating combination identifiers

Must be a single character

Default is "-"

nest_type

Output nesting format

'dt': Returns nested 'data table' (default)

'df': Returns nested 'data frame'

Value

'data table' containing nested transformation results

Includes 'pairs' column identifying column combinations

Contains 'data' column storing nested data structures

Supports optional grouping variables

Details

Advanced Transformation Mechanism:

Input validation and preprocessing

Dynamic column combination generation

Flexible pair transformation

Nested data structure creation

Transformation Process:

Validate input parameters

Generate column combinations

Create subset data tables

Merge and nest transformed data

Combination Strategy:

Supports variable-length column combinations

Generates all possible column pair permutations

Maintains original data context

Note

Key Operation Constraints:

Requires non-empty input data

cols2bind must specify existing columns

Supports flexible combination strategies

Computational complexity increases with combination size

See also

Examples

col_names <- c("Sepal.Length", "Sepal.Width", "Petal.Length")
c2p_nest(iris, cols2bind = col_names, pairs_n = 2, sep = "&")
#>                        pairs                data
#>                       <char>              <list>
#> 1:  Sepal.Length&Sepal.Width <data.table[150x4]>
#> 2: Sepal.Length&Petal.Length <data.table[150x4]>
#> 3:  Sepal.Width&Petal.Length <data.table[150x4]>
c2p_nest(iris, cols2bind = col_names, pairs_n = 2, by = "Species")
#>                        pairs    Species               data
#>                       <char>     <fctr>             <list>
#> 1:  Sepal.Length-Sepal.Width     setosa <data.table[50x3]>
#> 2:  Sepal.Length-Sepal.Width versicolor <data.table[50x3]>
#> 3:  Sepal.Length-Sepal.Width  virginica <data.table[50x3]>
#> 4: Sepal.Length-Petal.Length     setosa <data.table[50x3]>
#> 5: Sepal.Length-Petal.Length versicolor <data.table[50x3]>
#> 6: Sepal.Length-Petal.Length  virginica <data.table[50x3]>
#> 7:  Sepal.Width-Petal.Length     setosa <data.table[50x3]>
#> 8:  Sepal.Width-Petal.Length versicolor <data.table[50x3]>
#> 9:  Sepal.Width-Petal.Length  virginica <data.table[50x3]>