Generate and Combine Pairs of Columns and Nest Data

The combn_pair function generates combinations of specified columns, combines them, and nests the data by the specified columns. It supports both data.table and data.frame inputs and allows for optional grouping.

combn_pair(data, cols_bind, by = NULL, pairs_n = 2, sep = "-")

Arguments

data

data.table or data.frame. The input data.

cols_bind

character vector. The columns to combine into pairs.

by

character vector, optional. Columns to group by. Default is NULL.

pairs_n

integer, optional. The number of columns to include in each combination. Default is 2.

sep

character, optional. The separator to use when combining column names. Default is "-".

Value

data.table. A nested data.table with combined column pairs.

If by is NULL, returns a data.table nested by 'pairs'.

If by is specified, returns a data.table nested by 'pairs' and the grouping variables.

Note

The cols_bind parameter must be a character vector of column names.

The pairs_n parameter must be a positive integer greater than or equal to 2.

The sep parameter must be a single character string.

Ensure all grouping variables specified in by are present in the data.

Examples

col_names <- c("Sepal.Length", "Sepal.Width", "Petal.Length")
combn_pair(iris, cols_bind = col_names, pairs_n = 2, sep = "&")
#>                        pairs                data
#>                       <char>              <list>
#> 1:  Sepal.Length&Sepal.Width <data.table[150x4]>
#> 2: Sepal.Length&Petal.Length <data.table[150x4]>
#> 3:  Sepal.Width&Petal.Length <data.table[150x4]>
combn_pair(iris, cols_bind = col_names, pairs_n = 2, by = "Species")
#>                        pairs    Species               data
#>                       <char>     <fctr>             <list>
#> 1:  Sepal.Length-Sepal.Width     setosa <data.table[50x3]>
#> 2:  Sepal.Length-Sepal.Width versicolor <data.table[50x3]>
#> 3:  Sepal.Length-Sepal.Width  virginica <data.table[50x3]>
#> 4: Sepal.Length-Petal.Length     setosa <data.table[50x3]>
#> 5: Sepal.Length-Petal.Length versicolor <data.table[50x3]>
#> 6: Sepal.Length-Petal.Length  virginica <data.table[50x3]>
#> 7:  Sepal.Width-Petal.Length     setosa <data.table[50x3]>
#> 8:  Sepal.Width-Petal.Length versicolor <data.table[50x3]>
#> 9:  Sepal.Width-Petal.Length  virginica <data.table[50x3]>