Import CSV Files Using Specified Package

The import_csv function imports one or multiple CSV files using a specified package (data.table, vroom, or arrow). It supports row binding of multiple files into a single dataset with optional labeling of the origin file.

Usage

import_csv(file, package = "data.table", rbind = TRUE, rbind_label = NULL, ...)

Arguments

file

character vector of file paths. The paths to the CSV files to be imported.. These paths must point to existing files.

package

character, optional. Specifies which package to use for reading the CSV files. Default is 'data.table'. Valid options are data.table, vroom, and arrow.

rbind

logical, optional. If TRUE and more than one file is specified, it binds rows from different files into one dataset. Default is TRUE. Note: For vroom, row binding is inherent and cannot be turned off.

rbind_label

character, optional. The name of the column to add to the dataset that indicates the file origin. This is useful when row binding is used. Default is NULL.

...

Additional arguments passed to import_csv().The arguments vary depending on the R package chosen.

Value

Depending on the input and settings, returns a data.table or a list of data.tables. If rbind is TRUE and more than one file is processed, it returns a single data.table with all files combined. If rbind is FALSE or only one file is provided, it returns the data table from the single file processed.

Note

The function stops and throws an error if any file does not exist or if an unsupported package is specified.

The vroom package inherently binds rows when reading multiple files and does not support turning off this feature.

See also

Examples

csv_files <- mintyr_example(mintyr_examples("csv_test"))
csv_files
#> [1] "/home/runner/work/_temp/Library/mintyr/extdata/csv_test1.csv"
#> [2] "/home/runner/work/_temp/Library/mintyr/extdata/csv_test2.csv"
import_csv(csv_files)
#>     col1   col2   col3
#>    <int> <char> <lgcl>
#> 1:     4      d  FALSE
#> 2:     5      f   TRUE
#> 3:     6      e   TRUE
#> 4:    15      o  FALSE
#> 5:    16      p   TRUE
#> 6:    17      q  FALSE