Import list of xlsx sheets

The import_xlsx function reads data from Excel files, optionally from specified sheets, and can either merge data from multiple files or keep them separate.

import_xlsx(file, rbind = TRUE, sheet = NULL, ...)

Arguments

file

character vector of file paths. Paths to the Excel files to be imported. These paths must point to existing files.

rbind

logical, optional. If TRUE and more than one file is specified, it binds rows from different files into one dataset. Default is TRUE. Note: For vroom, row binding is inherent and cannot be turned off.

sheet

numeric vector or NULL, optional. Specifies the sheets to be read from each file. If NULL, all sheets are read. If numeric, only the specified sheets (by index) are read. Default is NULL.

...

Additional arguments to be passed to the readxl::read_excel() function, such as col_types to specify column data types.

Value

Depending on the rbind parameter:

If rbind is TRUE, returns a single data.table combining all files and sheets, with an additional 'excel_name' column indicating the source file.

If rbind is FALSE, returns a named list of data.tables, each representing data from a sheet, named by combining the file and sheet names.

Note

Make sure that the file paths provided are valid and exist. The function will stop with an error if any file does not exist or if the sheet indices are out of range.

Examples

xlsx_files <- mintyr_example(mintyr_examples("xlsx_test"))
xlsx_files
#> [1] "/home/runner/work/_temp/Library/mintyr/extdata/xlsx_test1.xlsx"
#> [2] "/home/runner/work/_temp/Library/mintyr/extdata/xlsx_test2.xlsx"
import_xlsx(xlsx_files)
#>     excel_name sheet_name  col1   col2   col3
#>         <char>     <char> <num> <char> <lgcl>
#>  1: xlsx_test1     Sheet1     4      d  FALSE
#>  2: xlsx_test1     Sheet1     5      f   TRUE
#>  3: xlsx_test1     Sheet1     6      e   TRUE
#>  4: xlsx_test1     Sheet2     1      a   TRUE
#>  5: xlsx_test1     Sheet2     2      b  FALSE
#>  6: xlsx_test1     Sheet2     3      c   TRUE
#>  7: xlsx_test2     Sheet1    15      o  FALSE
#>  8: xlsx_test2     Sheet1    16      p   TRUE
#>  9: xlsx_test2     Sheet1    17      q  FALSE
#> 10: xlsx_test2          a     7      g  FALSE
#> 11: xlsx_test2          a     9      h   TRUE
#> 12: xlsx_test2          a     8      i  FALSE
#> 13: xlsx_test2          b    10      J  FALSE
#> 14: xlsx_test2          b    11      K   TRUE
#> 15: xlsx_test2          b    12      L  FALSE