Apply Cross-Validation to Nested Data Tables

The nest_cv function performs cross-validation on nested data tables within a data.table. It supports various cross-validation configurations and handles nested lists of data tables.

nest_cv(
  nest_dt,
  v = 10,
  repeats = 1,
  strata = NULL,
  breaks = 4,
  pool = 0.1,
  ...
)

Arguments

nest_dt

data.table. The input data table containing nested data tables.

v

The number of partitions of the data set.

repeats

The number of times to repeat the V-fold partitioning.

strata

A variable in data (single character or name) used to conduct stratified sampling. When not NULL, each resample is created within the stratification variable. Numeric strata are binned into quartiles.

breaks

A single number giving the number of bins desired to stratify a numeric stratification variable.

pool

A proportion of data used to determine if a particular group is too small and should be pooled into another group. We do not recommend decreasing this argument below its default of 0.1 because of the dangers of stratifying groups that are too small.

...

These dots are for future extensions and must be empty.

Value

data.table. A data.table with cross-validation splits applied to the nested data tables.

Note

The nest_dt parameter must be a data.table containing at least one nested list column of data tables.

The function will copy the input data.table to avoid modifying the original data.

Examples

dt_nest <- w2l_nest(data = iris, cols2l = 1:2, by = "Species")
nest_cv(nest_dt = dt_nest, v = 2, repeats = 2)
#>             name    Species                   splits     id              train
#>           <fctr>     <fctr>                   <list> <char>             <list>
#>  1: Sepal.Length     setosa <vfold_split[45x5x50x3]> Fold01 <data.table[45x3]>
#>  2: Sepal.Length     setosa <vfold_split[45x5x50x3]> Fold02 <data.table[45x3]>
#>  3: Sepal.Length     setosa <vfold_split[45x5x50x3]> Fold03 <data.table[45x3]>
#>  4: Sepal.Length     setosa <vfold_split[45x5x50x3]> Fold04 <data.table[45x3]>
#>  5: Sepal.Length     setosa <vfold_split[45x5x50x3]> Fold05 <data.table[45x3]>
#>  6: Sepal.Length     setosa <vfold_split[45x5x50x3]> Fold06 <data.table[45x3]>
#>  7: Sepal.Length     setosa <vfold_split[45x5x50x3]> Fold07 <data.table[45x3]>
#>  8: Sepal.Length     setosa <vfold_split[45x5x50x3]> Fold08 <data.table[45x3]>
#>  9: Sepal.Length     setosa <vfold_split[45x5x50x3]> Fold09 <data.table[45x3]>
#> 10: Sepal.Length     setosa <vfold_split[45x5x50x3]> Fold10 <data.table[45x3]>
#> 11: Sepal.Length versicolor <vfold_split[45x5x50x3]> Fold01 <data.table[45x3]>
#> 12: Sepal.Length versicolor <vfold_split[45x5x50x3]> Fold02 <data.table[45x3]>
#> 13: Sepal.Length versicolor <vfold_split[45x5x50x3]> Fold03 <data.table[45x3]>
#> 14: Sepal.Length versicolor <vfold_split[45x5x50x3]> Fold04 <data.table[45x3]>
#> 15: Sepal.Length versicolor <vfold_split[45x5x50x3]> Fold05 <data.table[45x3]>
#> 16: Sepal.Length versicolor <vfold_split[45x5x50x3]> Fold06 <data.table[45x3]>
#> 17: Sepal.Length versicolor <vfold_split[45x5x50x3]> Fold07 <data.table[45x3]>
#> 18: Sepal.Length versicolor <vfold_split[45x5x50x3]> Fold08 <data.table[45x3]>
#> 19: Sepal.Length versicolor <vfold_split[45x5x50x3]> Fold09 <data.table[45x3]>
#> 20: Sepal.Length versicolor <vfold_split[45x5x50x3]> Fold10 <data.table[45x3]>
#> 21: Sepal.Length  virginica <vfold_split[45x5x50x3]> Fold01 <data.table[45x3]>
#> 22: Sepal.Length  virginica <vfold_split[45x5x50x3]> Fold02 <data.table[45x3]>
#> 23: Sepal.Length  virginica <vfold_split[45x5x50x3]> Fold03 <data.table[45x3]>
#> 24: Sepal.Length  virginica <vfold_split[45x5x50x3]> Fold04 <data.table[45x3]>
#> 25: Sepal.Length  virginica <vfold_split[45x5x50x3]> Fold05 <data.table[45x3]>
#> 26: Sepal.Length  virginica <vfold_split[45x5x50x3]> Fold06 <data.table[45x3]>
#> 27: Sepal.Length  virginica <vfold_split[45x5x50x3]> Fold07 <data.table[45x3]>
#> 28: Sepal.Length  virginica <vfold_split[45x5x50x3]> Fold08 <data.table[45x3]>
#> 29: Sepal.Length  virginica <vfold_split[45x5x50x3]> Fold09 <data.table[45x3]>
#> 30: Sepal.Length  virginica <vfold_split[45x5x50x3]> Fold10 <data.table[45x3]>
#> 31:  Sepal.Width     setosa <vfold_split[45x5x50x3]> Fold01 <data.table[45x3]>
#> 32:  Sepal.Width     setosa <vfold_split[45x5x50x3]> Fold02 <data.table[45x3]>
#> 33:  Sepal.Width     setosa <vfold_split[45x5x50x3]> Fold03 <data.table[45x3]>
#> 34:  Sepal.Width     setosa <vfold_split[45x5x50x3]> Fold04 <data.table[45x3]>
#> 35:  Sepal.Width     setosa <vfold_split[45x5x50x3]> Fold05 <data.table[45x3]>
#> 36:  Sepal.Width     setosa <vfold_split[45x5x50x3]> Fold06 <data.table[45x3]>
#> 37:  Sepal.Width     setosa <vfold_split[45x5x50x3]> Fold07 <data.table[45x3]>
#> 38:  Sepal.Width     setosa <vfold_split[45x5x50x3]> Fold08 <data.table[45x3]>
#> 39:  Sepal.Width     setosa <vfold_split[45x5x50x3]> Fold09 <data.table[45x3]>
#> 40:  Sepal.Width     setosa <vfold_split[45x5x50x3]> Fold10 <data.table[45x3]>
#> 41:  Sepal.Width versicolor <vfold_split[45x5x50x3]> Fold01 <data.table[45x3]>
#> 42:  Sepal.Width versicolor <vfold_split[45x5x50x3]> Fold02 <data.table[45x3]>
#> 43:  Sepal.Width versicolor <vfold_split[45x5x50x3]> Fold03 <data.table[45x3]>
#> 44:  Sepal.Width versicolor <vfold_split[45x5x50x3]> Fold04 <data.table[45x3]>
#> 45:  Sepal.Width versicolor <vfold_split[45x5x50x3]> Fold05 <data.table[45x3]>
#> 46:  Sepal.Width versicolor <vfold_split[45x5x50x3]> Fold06 <data.table[45x3]>
#> 47:  Sepal.Width versicolor <vfold_split[45x5x50x3]> Fold07 <data.table[45x3]>
#> 48:  Sepal.Width versicolor <vfold_split[45x5x50x3]> Fold08 <data.table[45x3]>
#> 49:  Sepal.Width versicolor <vfold_split[45x5x50x3]> Fold09 <data.table[45x3]>
#> 50:  Sepal.Width versicolor <vfold_split[45x5x50x3]> Fold10 <data.table[45x3]>
#> 51:  Sepal.Width  virginica <vfold_split[45x5x50x3]> Fold01 <data.table[45x3]>
#> 52:  Sepal.Width  virginica <vfold_split[45x5x50x3]> Fold02 <data.table[45x3]>
#> 53:  Sepal.Width  virginica <vfold_split[45x5x50x3]> Fold03 <data.table[45x3]>
#> 54:  Sepal.Width  virginica <vfold_split[45x5x50x3]> Fold04 <data.table[45x3]>
#> 55:  Sepal.Width  virginica <vfold_split[45x5x50x3]> Fold05 <data.table[45x3]>
#> 56:  Sepal.Width  virginica <vfold_split[45x5x50x3]> Fold06 <data.table[45x3]>
#> 57:  Sepal.Width  virginica <vfold_split[45x5x50x3]> Fold07 <data.table[45x3]>
#> 58:  Sepal.Width  virginica <vfold_split[45x5x50x3]> Fold08 <data.table[45x3]>
#> 59:  Sepal.Width  virginica <vfold_split[45x5x50x3]> Fold09 <data.table[45x3]>
#> 60:  Sepal.Width  virginica <vfold_split[45x5x50x3]> Fold10 <data.table[45x3]>
#>             name    Species                   splits     id              train
#>              validate
#>                <list>
#>  1: <data.table[5x3]>
#>  2: <data.table[5x3]>
#>  3: <data.table[5x3]>
#>  4: <data.table[5x3]>
#>  5: <data.table[5x3]>
#>  6: <data.table[5x3]>
#>  7: <data.table[5x3]>
#>  8: <data.table[5x3]>
#>  9: <data.table[5x3]>
#> 10: <data.table[5x3]>
#> 11: <data.table[5x3]>
#> 12: <data.table[5x3]>
#> 13: <data.table[5x3]>
#> 14: <data.table[5x3]>
#> 15: <data.table[5x3]>
#> 16: <data.table[5x3]>
#> 17: <data.table[5x3]>
#> 18: <data.table[5x3]>
#> 19: <data.table[5x3]>
#> 20: <data.table[5x3]>
#> 21: <data.table[5x3]>
#> 22: <data.table[5x3]>
#> 23: <data.table[5x3]>
#> 24: <data.table[5x3]>
#> 25: <data.table[5x3]>
#> 26: <data.table[5x3]>
#> 27: <data.table[5x3]>
#> 28: <data.table[5x3]>
#> 29: <data.table[5x3]>
#> 30: <data.table[5x3]>
#> 31: <data.table[5x3]>
#> 32: <data.table[5x3]>
#> 33: <data.table[5x3]>
#> 34: <data.table[5x3]>
#> 35: <data.table[5x3]>
#> 36: <data.table[5x3]>
#> 37: <data.table[5x3]>
#> 38: <data.table[5x3]>
#> 39: <data.table[5x3]>
#> 40: <data.table[5x3]>
#> 41: <data.table[5x3]>
#> 42: <data.table[5x3]>
#> 43: <data.table[5x3]>
#> 44: <data.table[5x3]>
#> 45: <data.table[5x3]>
#> 46: <data.table[5x3]>
#> 47: <data.table[5x3]>
#> 48: <data.table[5x3]>
#> 49: <data.table[5x3]>
#> 50: <data.table[5x3]>
#> 51: <data.table[5x3]>
#> 52: <data.table[5x3]>
#> 53: <data.table[5x3]>
#> 54: <data.table[5x3]>
#> 55: <data.table[5x3]>
#> 56: <data.table[5x3]>
#> 57: <data.table[5x3]>
#> 58: <data.table[5x3]>
#> 59: <data.table[5x3]>
#> 60: <data.table[5x3]>
#>              validate