Select Top Percentage of Data and Compute Summary Statistics

The top_perc function selects the top percentage of data based on a specified trait and computes summary statistics. It allows for grouping by additional columns and offers flexibility in the type of statistics calculated. The function can also retain the selected data if needed.

top_perc(data, perc, trait, by = NULL, type = "mean_sd", keep_data = FALSE)

Arguments

data

data.frame. The input data frame.

perc

numeric vector. A vector of percentages to select the top data. Each element should be a numeric value between -1 and 1.

trait

character. The name of the column in data on which to base the selection.

by

character vector, optional. The names of the columns to group by. Default is NULL.

type

type of summary statistics. Possible values include: "full", "common", "robust", "five_number", "mean_sd", "mean_se", "mean_ci", "median_iqr", "median_mad", "quantile", "mean", "median", "min", "max"

keep_data

logical, optional. If TRUE, the function returns both the summary statistics and the selected top data. Default is FALSE.

Value

A list or data frame:

If keep_data is FALSE, a data frame with summary statistics.

If keep_data is TRUE, a list where each element is a list containing summary statistics (stat) and the selected top data (data).

Note

The perc parameter accepts values between -1 and 1. Positive values select the top percentage, while negative values select the bottom percentage.

The function performs initial checks to ensure required arguments are provided and valid.

Grouping by additional columns (by) is optional and allows for more granular analysis.

The type parameter specifies the type of summary statistics to compute, with "mean_sd" as the default.

If keep_data is set to TRUE, the function will return both the summary statistics and the selected top data for each percentage.

Examples

top_perc(iris, perc = 0.5, trait = "Petal.Width", by = "Species")
#> # A tibble: 3 × 6
#>   Species    variable        n  mean    sd top_perc
#>   <fct>      <fct>       <dbl> <dbl> <dbl> <chr>   
#> 1 setosa     Petal.Width    45 0.262 0.098 50%     
#> 2 versicolor Petal.Width    35 1.43  0.13  50%     
#> 3 virginica  Petal.Width    29 2.22  0.165 50%