Reshape Wide Data to Long and Nest by Specified Columns

The w2l_nest function reshapes wide-format data to long-format and nests it by specified columns. It is designed to handle both data.frame and data.table objects and provides options for grouping the data.

w2l_nest(data, cols2l, by = NULL)

Arguments

data

data.frame or data.table. The input data in wide format.

cols2l

numeric or character vector. Specifies the columns to reshape from wide to long format. Can be either numeric indices or column names.

by

character vector, optional. Specifies the columns to group by. Default is NULL.

Value

data.table. A data.table with nested data in long format, grouped by specified columns if provided.

If by is NULL, returns a data.table nested by 'name'.

If by is specified, returns a data.table nested by 'name' and the grouping variables.

Note

The cols2l parameter should be either numeric indices or a character vector of column names.

Ensure all grouping variables specified in by are present in the data.

The function converts data.frame to data.table if necessary.

Examples

w2l_nest(data = iris, cols2l = 1:4)
#>            name                data
#>          <fctr>              <list>
#> 1: Sepal.Length <data.table[150x2]>
#> 2:  Sepal.Width <data.table[150x2]>
#> 3: Petal.Length <data.table[150x2]>
#> 4:  Petal.Width <data.table[150x2]>
w2l_nest(data = iris, cols2l = c("Sepal.Length", "Sepal.Width", "Petal.Length"), by = "Species")
#>            name    Species               data
#>          <fctr>     <fctr>             <list>
#> 1: Sepal.Length     setosa <data.table[50x2]>
#> 2: Sepal.Length versicolor <data.table[50x2]>
#> 3: Sepal.Length  virginica <data.table[50x2]>
#> 4:  Sepal.Width     setosa <data.table[50x2]>
#> 5:  Sepal.Width versicolor <data.table[50x2]>
#> 6:  Sepal.Width  virginica <data.table[50x2]>
#> 7: Petal.Length     setosa <data.table[50x2]>
#> 8: Petal.Length versicolor <data.table[50x2]>
#> 9: Petal.Length  virginica <data.table[50x2]>