Reshape Wide Data to Long Format and Nest by Specified Columns

The w2l_nest function reshapes wide-format data into long-format and nests it by specified columns. It handles both data.frame and data.table objects and provides options for grouping and nesting the data.

w2l_nest(data, cols2l, by = NULL, nest_type = "dt")

Arguments

data

A data.frame or data.table. The input data in wide format.

cols2l

A numeric or character vector. Specifies the columns to reshape from wide to long format. Can be either numeric indices or column names.

by

An optional character vector. Specifies the columns to group by. Default is NULL.

nest_type

A character string, either "dt" or "df". Specifies the type of object to nest in the result: "dt" for data.table, "df" for data.frame. Default is "dt".

Value

A data.table with nested data in long format, grouped by specified columns if provided. Each row contains a nested data.table or data.frame under the column data, depending on nest_type.

If by is NULL, returns a data.table nested by name.

If by is specified, returns a data.table nested by name and the grouping variables.

Details

The function performs the following key operations:

Melts the specified wide columns into long format

Nests the resulting data by the name column

Optionally groups by additional variables specified in by

Allows flexible output format (data.table or data.frame)

Transformation Strategy:

Converts input to data.table if necessary

Validates input columns and grouping variables

Reshapes data using data.table::melt()

Nests data with flexible grouping options

Note

Important Considerations:

The cols2l parameter should be either numeric indices or a character vector of column names present in data.

Ensure all grouping variables specified in by are present in data.

The function converts data.frame to data.table if necessary.

The nest_type parameter controls whether nested data are data.table ("dt") or data.frame ("df") objects.

If nest_type is not "dt" or "df", the function will stop with an error.

Examples

w2l_nest(data = iris, cols2l = 1:4)
#>            name                data
#>          <fctr>              <list>
#> 1: Sepal.Length <data.table[150x2]>
#> 2:  Sepal.Width <data.table[150x2]>
#> 3: Petal.Length <data.table[150x2]>
#> 4:  Petal.Width <data.table[150x2]>
w2l_nest(data = iris, cols2l = c("Sepal.Length", "Sepal.Width", "Petal.Length"), by = "Species")
#>            name    Species               data
#>          <fctr>     <fctr>             <list>
#> 1: Sepal.Length     setosa <data.table[50x2]>
#> 2: Sepal.Length versicolor <data.table[50x2]>
#> 3: Sepal.Length  virginica <data.table[50x2]>
#> 4:  Sepal.Width     setosa <data.table[50x2]>
#> 5:  Sepal.Width versicolor <data.table[50x2]>
#> 6:  Sepal.Width  virginica <data.table[50x2]>
#> 7: Petal.Length     setosa <data.table[50x2]>
#> 8: Petal.Length versicolor <data.table[50x2]>
#> 9: Petal.Length  virginica <data.table[50x2]>