Reshape Wide Data to Long Format and Split into List

The w2l_split function reshapes wide-format data into long-format and splits it into a list by variable names and optional grouping columns. It handles both data.frame and data.table objects.

w2l_split(data, cols2l, by = NULL, split_type = "dt")

Arguments

data

A data.frame or data.table. The input data in wide format.

cols2l

A numeric or character vector. Specifies the columns to reshape from wide to long format. Can be either numeric indices or column names.

by

An optional character vector. Specifies additional columns to split by. Default is NULL.

split_type

A character string, either "dt" or "df". Specifies the type of objects in the resulting list: "dt" for data.table, "df" for data.frame. Default is "dt".

Value

A list of data.table or data.frame objects (depending on split_type), split by variable names and optional grouping columns.

If by is NULL, returns a list split by variable names only.

If by is specified, returns a list split by both variable names and grouping variables.

Details

The function first melts the specified wide columns into long format and then splits the resulting data into a list based on the variable names and any additional grouping variables specified in by. The resulting list elements can be either data.table or data.frame objects, as specified by the split_type parameter.

Note

The cols2l parameter should be either numeric indices or a character vector of column names present in data.

Ensure all grouping variables specified in by are present in data.

The function automatically converts data.frame to data.table for processing.

The split_type parameter determines the class of objects in the resulting list.

Non-specified columns are automatically used as ID variables in the reshaping process.

Examples

w2l_split(data = iris, cols2l = 1:3)
#> $Sepal.Length
#>      Petal.Width   Species value
#>            <num>    <fctr> <num>
#>   1:         0.2    setosa   5.1
#>   2:         0.2    setosa   4.9
#>   3:         0.2    setosa   4.7
#>   4:         0.2    setosa   4.6
#>   5:         0.2    setosa   5.0
#>  ---                            
#> 146:         2.3 virginica   6.7
#> 147:         1.9 virginica   6.3
#> 148:         2.0 virginica   6.5
#> 149:         2.3 virginica   6.2
#> 150:         1.8 virginica   5.9
#> 
#> $Sepal.Width
#>      Petal.Width   Species value
#>            <num>    <fctr> <num>
#>   1:         0.2    setosa   3.5
#>   2:         0.2    setosa   3.0
#>   3:         0.2    setosa   3.2
#>   4:         0.2    setosa   3.1
#>   5:         0.2    setosa   3.6
#>  ---                            
#> 146:         2.3 virginica   3.0
#> 147:         1.9 virginica   2.5
#> 148:         2.0 virginica   3.0
#> 149:         2.3 virginica   3.4
#> 150:         1.8 virginica   3.0
#> 
#> $Petal.Length
#>      Petal.Width   Species value
#>            <num>    <fctr> <num>
#>   1:         0.2    setosa   1.4
#>   2:         0.2    setosa   1.4
#>   3:         0.2    setosa   1.3
#>   4:         0.2    setosa   1.5
#>   5:         0.2    setosa   1.4
#>  ---                            
#> 146:         2.3 virginica   5.2
#> 147:         1.9 virginica   5.0
#> 148:         2.0 virginica   5.2
#> 149:         2.3 virginica   5.4
#> 150:         1.8 virginica   5.1
#> 
w2l_split(data = iris, cols2l = c("Sepal.Length", "Sepal.Width"))
#> $Sepal.Length
#>      Petal.Length Petal.Width   Species value
#>             <num>       <num>    <fctr> <num>
#>   1:          1.4         0.2    setosa   5.1
#>   2:          1.4         0.2    setosa   4.9
#>   3:          1.3         0.2    setosa   4.7
#>   4:          1.5         0.2    setosa   4.6
#>   5:          1.4         0.2    setosa   5.0
#>  ---                                         
#> 146:          5.2         2.3 virginica   6.7
#> 147:          5.0         1.9 virginica   6.3
#> 148:          5.2         2.0 virginica   6.5
#> 149:          5.4         2.3 virginica   6.2
#> 150:          5.1         1.8 virginica   5.9
#> 
#> $Sepal.Width
#>      Petal.Length Petal.Width   Species value
#>             <num>       <num>    <fctr> <num>
#>   1:          1.4         0.2    setosa   3.5
#>   2:          1.4         0.2    setosa   3.0
#>   3:          1.3         0.2    setosa   3.2
#>   4:          1.5         0.2    setosa   3.1
#>   5:          1.4         0.2    setosa   3.6
#>  ---                                         
#> 146:          5.2         2.3 virginica   3.0
#> 147:          5.0         1.9 virginica   2.5
#> 148:          5.2         2.0 virginica   3.0
#> 149:          5.4         2.3 virginica   3.4
#> 150:          5.1         1.8 virginica   3.0
#>