Wide to Long Data Splitting with Advanced Transformation Capabilities

Transforms wide-format data to long format with flexible splitting and grouping options, supporting multiple data type conversions.

w2l_split(data, cols2l, by = NULL, split_type = "dt", sep = "_")

Arguments

data

A data.frame or data.table to be transformed and split.

cols2l

Columns to be transformed from wide to long format. Can be specified as numeric indices or column names.

by

Optional grouping columns for additional data stratification.

split_type

Output format for split data, either "dt" (data.table) or "df" (data.frame). Defaults to "dt".

sep

Separator used for generating list names when grouping. Defaults to "_".

Value

A list of data.tables or data.frames, split according to specified parameters.

Details

The function provides comprehensive wide-to-long data transformation with the following advanced features:

Flexible column specification (numeric indices or names)

Automatic data type conversion

Intelligent handling of grouping variables

Customizable output data type

Robust error checking and validation

Transformation Strategy:

Converts input to data.table format

Melts specified columns to long format

Splits data based on optional grouping variables

Generates list names using grouping values

Column Specification Options:

Numeric indices of columns to transform

Character vector of column names

Comprehensive validation of provided columns

Note

Supports complex data transformation scenarios

Handles mixed input data types

Provides detailed error messages for invalid inputs

Efficient data manipulation using data.table

Examples

w2l_split(data = iris, cols2l = 1:3)
#> $Sepal.Length
#>      Petal.Width   Species value
#>            <num>    <fctr> <num>
#>   1:         0.2    setosa   5.1
#>   2:         0.2    setosa   4.9
#>   3:         0.2    setosa   4.7
#>   4:         0.2    setosa   4.6
#>   5:         0.2    setosa   5.0
#>  ---                            
#> 146:         2.3 virginica   6.7
#> 147:         1.9 virginica   6.3
#> 148:         2.0 virginica   6.5
#> 149:         2.3 virginica   6.2
#> 150:         1.8 virginica   5.9
#> 
#> $Sepal.Width
#>      Petal.Width   Species value
#>            <num>    <fctr> <num>
#>   1:         0.2    setosa   3.5
#>   2:         0.2    setosa   3.0
#>   3:         0.2    setosa   3.2
#>   4:         0.2    setosa   3.1
#>   5:         0.2    setosa   3.6
#>  ---                            
#> 146:         2.3 virginica   3.0
#> 147:         1.9 virginica   2.5
#> 148:         2.0 virginica   3.0
#> 149:         2.3 virginica   3.4
#> 150:         1.8 virginica   3.0
#> 
#> $Petal.Length
#>      Petal.Width   Species value
#>            <num>    <fctr> <num>
#>   1:         0.2    setosa   1.4
#>   2:         0.2    setosa   1.4
#>   3:         0.2    setosa   1.3
#>   4:         0.2    setosa   1.5
#>   5:         0.2    setosa   1.4
#>  ---                            
#> 146:         2.3 virginica   5.2
#> 147:         1.9 virginica   5.0
#> 148:         2.0 virginica   5.2
#> 149:         2.3 virginica   5.4
#> 150:         1.8 virginica   5.1
#> 
w2l_split(data = iris, cols2l = c("Sepal.Length", "Sepal.Width"))
#> $Sepal.Length
#>      Petal.Length Petal.Width   Species value
#>             <num>       <num>    <fctr> <num>
#>   1:          1.4         0.2    setosa   5.1
#>   2:          1.4         0.2    setosa   4.9
#>   3:          1.3         0.2    setosa   4.7
#>   4:          1.5         0.2    setosa   4.6
#>   5:          1.4         0.2    setosa   5.0
#>  ---                                         
#> 146:          5.2         2.3 virginica   6.7
#> 147:          5.0         1.9 virginica   6.3
#> 148:          5.2         2.0 virginica   6.5
#> 149:          5.4         2.3 virginica   6.2
#> 150:          5.1         1.8 virginica   5.9
#> 
#> $Sepal.Width
#>      Petal.Length Petal.Width   Species value
#>             <num>       <num>    <fctr> <num>
#>   1:          1.4         0.2    setosa   3.5
#>   2:          1.4         0.2    setosa   3.0
#>   3:          1.3         0.2    setosa   3.2
#>   4:          1.5         0.2    setosa   3.1
#>   5:          1.4         0.2    setosa   3.6
#>  ---                                         
#> 146:          5.2         2.3 virginica   3.0
#> 147:          5.0         1.9 virginica   2.5
#> 148:          5.2         2.0 virginica   3.0
#> 149:          5.4         2.3 virginica   3.4
#> 150:          5.1         1.8 virginica   3.0
#>