Convert Wide Data to Long Format and Split into List of Data Tables

The w2l_split function reshapes wide-format data to long-format and splits it into a list of data tables, each corresponding to one of the specified columns.

w2l_split(data, cols2l)

Arguments

data

data.frame or data.table. The input data in wide format.

cols2l

numeric or character vector. Specifies the columns to reshape from wide to long format. Can be either numeric indices or column names.

Value

A list of data tables. Each data table in the list corresponds to one of the specified columns, reshaped into long format.

Note

The cols2l parameter should be either numeric indices or a character vector of column names.

Ensure all specified columns in cols2l are present in the data.

The function converts data.frame to data.table if necessary.

Examples

w2l_split(data = iris, cols2l = 1:3)
#> $Sepal.Length
#>      Petal.Width   Species     variable value
#>            <num>    <fctr>       <fctr> <num>
#>   1:         0.2    setosa Sepal.Length   5.1
#>   2:         0.2    setosa Sepal.Length   4.9
#>   3:         0.2    setosa Sepal.Length   4.7
#>   4:         0.2    setosa Sepal.Length   4.6
#>   5:         0.2    setosa Sepal.Length   5.0
#>  ---                                         
#> 146:         2.3 virginica Sepal.Length   6.7
#> 147:         1.9 virginica Sepal.Length   6.3
#> 148:         2.0 virginica Sepal.Length   6.5
#> 149:         2.3 virginica Sepal.Length   6.2
#> 150:         1.8 virginica Sepal.Length   5.9
#> 
#> $Sepal.Width
#>      Petal.Width   Species    variable value
#>            <num>    <fctr>      <fctr> <num>
#>   1:         0.2    setosa Sepal.Width   3.5
#>   2:         0.2    setosa Sepal.Width   3.0
#>   3:         0.2    setosa Sepal.Width   3.2
#>   4:         0.2    setosa Sepal.Width   3.1
#>   5:         0.2    setosa Sepal.Width   3.6
#>  ---                                        
#> 146:         2.3 virginica Sepal.Width   3.0
#> 147:         1.9 virginica Sepal.Width   2.5
#> 148:         2.0 virginica Sepal.Width   3.0
#> 149:         2.3 virginica Sepal.Width   3.4
#> 150:         1.8 virginica Sepal.Width   3.0
#> 
#> $Petal.Length
#>      Petal.Width   Species     variable value
#>            <num>    <fctr>       <fctr> <num>
#>   1:         0.2    setosa Petal.Length   1.4
#>   2:         0.2    setosa Petal.Length   1.4
#>   3:         0.2    setosa Petal.Length   1.3
#>   4:         0.2    setosa Petal.Length   1.5
#>   5:         0.2    setosa Petal.Length   1.4
#>  ---                                         
#> 146:         2.3 virginica Petal.Length   5.2
#> 147:         1.9 virginica Petal.Length   5.0
#> 148:         2.0 virginica Petal.Length   5.2
#> 149:         2.3 virginica Petal.Length   5.4
#> 150:         1.8 virginica Petal.Length   5.1
#> 
w2l_split(data = iris, cols2l = c("Sepal.Length", "Sepal.Width"))
#> $Sepal.Length
#>      Petal.Length Petal.Width   Species     variable value
#>             <num>       <num>    <fctr>       <fctr> <num>
#>   1:          1.4         0.2    setosa Sepal.Length   5.1
#>   2:          1.4         0.2    setosa Sepal.Length   4.9
#>   3:          1.3         0.2    setosa Sepal.Length   4.7
#>   4:          1.5         0.2    setosa Sepal.Length   4.6
#>   5:          1.4         0.2    setosa Sepal.Length   5.0
#>  ---                                                      
#> 146:          5.2         2.3 virginica Sepal.Length   6.7
#> 147:          5.0         1.9 virginica Sepal.Length   6.3
#> 148:          5.2         2.0 virginica Sepal.Length   6.5
#> 149:          5.4         2.3 virginica Sepal.Length   6.2
#> 150:          5.1         1.8 virginica Sepal.Length   5.9
#> 
#> $Sepal.Width
#>      Petal.Length Petal.Width   Species    variable value
#>             <num>       <num>    <fctr>      <fctr> <num>
#>   1:          1.4         0.2    setosa Sepal.Width   3.5
#>   2:          1.4         0.2    setosa Sepal.Width   3.0
#>   3:          1.3         0.2    setosa Sepal.Width   3.2
#>   4:          1.5         0.2    setosa Sepal.Width   3.1
#>   5:          1.4         0.2    setosa Sepal.Width   3.6
#>  ---                                                     
#> 146:          5.2         2.3 virginica Sepal.Width   3.0
#> 147:          5.0         1.9 virginica Sepal.Width   2.5
#> 148:          5.2         2.0 virginica Sepal.Width   3.0
#> 149:          5.4         2.3 virginica Sepal.Width   3.4
#> 150:          5.1         1.8 virginica Sepal.Width   3.0
#>